Fixed queries according to LRB validator  - changed AVG compuation to double type  - changed car count to "count distinct"  - added special sinks to get correct attribute order in output  - changed unknown LAV value to "-1" Added FlushStream Added Subqueries and reworked query classes  - Subqueries for: AccDec, AvgVSpd, AvgSpd, Lav, VCnt  - reworked queries for: Accidents, Tolls, and compined both in LRB Additional code clean up

+++ queries/lrb/src/main/java/de/hub/cs/dbis/lrb/types/internal/AvgSpeedTuple.java
@@ -18,9 +18,9 @@
  */
 package de.hub.cs.dbis.lrb.types.internal;
 
-import storm.lrb.TopologyControl;
 import backtype.storm.tuple.Fields;
 import backtype.storm.tuple.Values;
+import de.hub.cs.dbis.lrb.queries.utils.TopologyControl;
 import de.hub.cs.dbis.lrb.types.util.ISegmentIdentifier;
 import de.hub.cs.dbis.lrb.util.Time;
 
@@ -80,7 +80,7 @@ public final class AvgSpeedTuple extends Values implements ISegmentIdentifier {
 	 * @param avgSpeed
 	 *            the average speed of the vehicle
 	 */
-	public AvgSpeedTuple(Short minute, Integer xway, Short segment, Short diretion, Integer avgSpeed) {
+	public AvgSpeedTuple(Short minute, Integer xway, Short segment, Short diretion, Double avgSpeed) {
 		assert (minute != null);
 		assert (xway != null);
 		assert (segment != null);
@@ -140,8 +140,8 @@ public final class AvgSpeedTuple extends Values implements ISegmentIdentifier {
 	 * 
 	 * @return the average speed of this tuple
 	 */
-	public final Integer getAvgSpeed() {
-		return (Integer)super.get(AVGS_IDX);
+	public final Double getAvgSpeed() {
+		return (Double)super.get(AVGS_IDX);
 	}
 	
 	/**
